SnowGPT_ChatBot/requirements.txt
langchain==0.0.268
openai==0.27.8
pinecone_client==2.2.2
sentence_transformers==2.2.2
streamlit==1.25.0
streamlit_chat==0.1.1
